Examples related to class03 with arrays
